pyproject.toml
[project]
name = "datetime-heatmap-notion"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ches", email = "chestnut31415@gmail.com" }
]
dependencies = [
    "seaborn>=0.13.2",
    "matplotlib>=3.9.2",
    "matplotlib-fontja>=1.0.0",
    "requests>=2.32.3",
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/datetime_heatmap_notion"]
